# packages/api-client/openapi.json
{
  "openapi": "3.1.1",
...
  "info": {
    "title": "UniVarse API",
    "version": "v1",
    "description": "Generated by `pnpm contracts:gen` from the API route metadata (spec 0011). Do not edit by hand."
  },
...
  "paths": {
...
    "/api/v1/settings": {
      "get": {
        "operationId": "listSettings",
        "summary": "Settings you can see, with their values",
        "tags": [
          "settings"
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        },
        "security": [
          {
            "session": []
          }
        ],
        "x-product": "core",
        "x-permission": "settings.tenant.view",
        "x-step-up": false
      }
    },
    "/api/v1/settings/{key}": {
      "get": {
        "operationId": "getSetting",
        "summary": "One setting, with its ETag",
        "tags": [
          "settings"
        ],
        "parameters": [
          {
            "name": "key",
            "in": "path",
            "required": true,
            "description": "A setting key from the registry. An unknown key, or one whose product is off, is 404 settings.unknown_key. Writing needs the key’s own manage permission (x-permission-by-key).",
            "schema": {
              "type": "string",
              "enum": [
                "registration.unitLimits",
                "files.storageQuotaBytes"
              ]
            },
            "x-invalid": {
              "status": 404,
              "code": "settings.unknown_key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        },
        "security": [
          {
            "session": []
          }
        ],
        "x-product": "core",
        "x-permission": "settings.tenant.view",
        "x-step-up": false
      },
      "put": {
        "operationId": "putSetting",
        "summary": "Change a setting (If-Match required); the value must fit the chosen key",
        "tags": [
          "settings"
        ],
        "parameters": [
          {
            "name": "key",
            "in": "path",
            "required": true,
            "description": "A setting key from the registry. An unknown key, or one whose product is off, is 404 settings.unknown_key. Writing needs the key’s own manage permission (x-permission-by-key).",
            "schema": {
              "type": "string",
              "enum": [
                "registration.unitLimits",
                "files.storageQuotaBytes"
              ]
            },
            "x-invalid": {
              "status": 404,
              "code": "settings.unknown_key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "value": {
                    "anyOf": [
                      {
                        "$ref": "#/components/schemas/SettingValue.registration.unitLimits"
                      },
                      {
                        "$ref": "#/components/schemas/SettingValue.files.storageQuotaBytes"
                      }
                    ]
                  }
                },
                "required": [
                  "value"
                ],
                "additionalProperties": false,
                "x-value-by-key": {
                  "parameter": "key",
                  "values": {
                    "registration.unitLimits": "#/components/schemas/SettingValue.registration.unitLimits",
                    "files.storageQuotaBytes": "#/components/schemas/SettingValue.files.storageQuotaBytes"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        },
        "security": [
          {
            "session": []
          }
        ],
        "x-product": "core",
        "x-permission": "settings.tenant.view",
        "x-permission-by-key": {
          "registration.unitLimits": {
            "permission": "settings.registration.manage",
            "stepUp": true
          },
          "files.storageQuotaBytes": {
            "permission": "settings.tenant.manage",
            "stepUp": true
          }
        },
        "x-step-up": false,
        "x-csrf": "same-origin"
      },
      "delete": {
        "operationId": "resetSetting",
        "summary": "Return a setting to its default (If-Match required)",
        "tags": [
          "settings"
        ],
        "parameters": [
          {
            "name": "key",
            "in": "path",
            "required": true,
            "description": "A setting key from the registry. An unknown key, or one whose product is off, is 404 settings.unknown_key. Writing needs the key’s own manage permission (x-permission-by-key).",
            "schema": {
              "type": "string",
              "enum": [
                "registration.unitLimits",
                "files.storageQuotaBytes"
              ]
            },
            "x-invalid": {
              "status": 404,
              "code": "settings.unknown_key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        },
        "security": [
          {
            "session": []
          }
        ],
        "x-product": "core",
        "x-permission": "settings.tenant.view",
        "x-permission-by-key": {
          "registration.unitLimits": {
            "permission": "settings.registration.manage",
            "stepUp": true
          },
          "files.storageQuotaBytes": {
            "permission": "settings.tenant.manage",
            "stepUp": true
          }
        },
        "x-step-up": false,
        "x-csrf": "same-origin"
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "session": {
        "type": "apiKey",
        "in": "cookie",
        "name": "__Host-uv_sid",
        "description": "Set by login, MFA verify and step-up; HttpOnly."
      },
      "mfaChallenge": {
        "type": "apiKey",
        "in": "cookie",
        "name": "__Host-uv_mfa",
        "description": "Set by login when a second factor is needed; read only by MFA verify."
      }
    },
    "schemas": {
      "SettingValue.files.storageQuotaBytes": {
        "title": "files.storageQuotaBytes",
        "type": "integer",
        "minimum": 0,
        "maximum": 10995116277760
      },
      "SettingValue.registration.unitLimits": {
        "title": "registration.unitLimits",
        "type": "object",
        "properties": {
          "min": {
            "type": "integer",
            "minimum": 1,
            "maximum": 60
          },
          "max": {
            "type": "integer",
            "minimum": 1,
            "maximum": 60
          }
        },
        "required": [
          "min",
          "max"
        ],
        "additionalProperties": false
      }
    }
  }
}